07/07/14 --> In Create Account Activity, adds option to either take a profile photo using the camera or upload one from the photo gallery. Allows cropping/resizing of photo to fit the photo square.
On LogIn Activity, disables the login button from being clicked again after it was already clicked. Also shows a loading log-in dialog.

Added AsyncTask for downloading images from a URL (passed as a parameter) and saves them to the SDCard with the date in milliseconds as the file name and a .PNG extension